Pr update welding demo (#46)
* Update renamed method fillMessage()->appendTo()

In commit 397fc070ea87594c0fc9a01b1196dd65b3064a4d of the
moveit_task_constructor repo, the `fillMessage` method of the
SolutionBase class was renamed to `appendTo`. This commit applies that
renaming to the demo code.

* Update welding demo urdf xacro

Update `ur_robot` macro params

Remove "robot_platform_joint" as the updated `ur_robot` macro takes care
of connecting the robot to the provided parent link now.

Use `xacro.load_yaml` instead of deprecated `load_yaml` function

NOTE: these changes are required if the demo is run with an updated
version of the ur_description package that provides the `ur_robot` macro

* Add warehouse_ros_mongo ros2 branch to repos

Also add to dependencies in package.xml

* Update controller manager executable name

spawner.py -> spawner

* Update visual_parameters.yaml structure

* Add required topic names to the local_planner conf

Otherwise these topics by default are set to "<undefined>" which causes
an error when the local planner node attempts to subscribe to them as
topic names are not allowed to have the '<' or '>' character in them.

* Update controller name

The old ur_controllers/ForceTorqueStateBroadcaster controller does not
exist anymore. The updated one to use is from the
force_torque_sensor_broadcaster package

* Formatting Updates

Update pre-commit config and fix formatting/spelling throughout repo

Update github formatting workflow

* Update ur_msgs repo to an existing ros2 branch

* Comment out warehouse_ros_mongo code

The warehouse_ros_mongo package cannot be used on Ubuntu Jammy. It
depends on mongodb which has no released version on Jammy. So comment
out all portions of the welding demo that reference it.

* More formatting changes

Remove extra spaces in yaml files to get CI passing

* Remove ShellCheck since there are no shell scripts

* Update PlanningSceneMonitor usage

Sending the TF buffer argument in the constructor of a
PlanningSceneMonitor object is deprecated.

* Add tinyxml_vendor and tinyxml dependency details

* Moveit_servo exec_depend->depend

moveit_servo is used in the ServoSolver c++ code so it is needed as a
build time dependency as well.

* Remove unneeded source repos

* CI: update to humble